@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "com.prolificinteractive.materialcalendarview.sample"
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode project.ext.versionCodeInt
         versionName version
     }
@@ -20,10 +20,10 @@ android {
 dependencies {
     // You should use the commented out line below in you're application.
     // We depend on the source directly here so that development is easier.
-    compile project(':library')
+    api project(':library')
     //compile 'com.prolificinteractive:material-calendarview:1.4.0'
 
-    compile 'com.android.support:appcompat-v7:25.1.1'
-    compile 'com.android.support:recyclerview-v7:25.1.1'
-    compile 'com.jakewharton:butterknife:7.0.1'
+    api 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    api 'androidx.recyclerview:recyclerview:1.1.0-alpha01'
+    api 'com.jakewharton:butterknife:8.8.1'
 }
