@@ -2,9 +2,6 @@ package com.prolificinteractive.materialcalendarview.sample;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,7 +10,10 @@ import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
 
 import java.util.Calendar;
 
-import butterknife.Bind;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import butterknife.BindView;
 import butterknife.ButterKnife;
 
 /**
@@ -25,7 +25,7 @@ public class MultipleViewActivity extends AppCompatActivity{
     //number of MaterialCalendarViews to display in list
     static final int NUM_ENTRIES = 3;
 
-    @Bind(R.id.calendar_list) RecyclerView calendarList;
+    @BindView(R.id.calendar_list) RecyclerView calendarList;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
